// src/app/api/contribuyente/models/documento-sustento-buscar-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DocumentoSustentoBuscarResponse {
  activo?: number;
  contribuyenteNumero?: number;
  docSusContribuyenteId?: number;
  estadoDsc?: string;
  folios?: number;
  forPresentacionId?: number;
  municipalidadId?: number;
  nroDocSustento?: string;
  rowNum?: number;
  tipDocSustentoId?: number;
  tipoDocumentoDsc?: string;
  tipoPresentacionDsc?: string;
}